refactor: rename folder + format code
