+++ front-end/mobile/lib/pages/Home/bot_message.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:mobile/pages/Home/typewriter_text.dart';
 
-
-
 class BotMessage extends StatelessWidget {
   final String message;
